pyproject.toml
[project]
name = "acme"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Neal Paresh Patel", email = "59032397+Desitrain22@users.noreply.github.com" }
]
dependencies = [
    "flask[dotenv]>=3.1.1",
    "jinja-partials>=0.3.0",
    "feedparser>=6.0.11",
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/acme"]
